% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_sample_cleaning.R
\name{epinion_data_cleaning}
\alias{epinion_data_cleaning}
\title{Clean data header before uploading sample to SPSSD}
\usage{
epinion_data_cleaning(data, output)
}
\arguments{
\item{data}{the name of the dataframe.}

\item{output}{name of the text file}
}
\value{
A cleaned dataframe
}
\description{
This function removes semi-colon, double-quote, tab and line break if detected anywhere in the data.
Remove special and non-ASCII character in data header.
Replace some Danish characters that may cause the error when uploading the sample to SPSSD with the Latin characters.
Save the cleaned data to a .txt file under working directory
}
\examples{
data_final <- epinion_data_cleaning(data_email_validated, "sample.txt")
}
