% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_sample_cleaning.R
\name{epinion_email_validation}
\alias{epinion_email_validation}
\title{Validate format of email field in the data}
\usage{
epinion_email_validation(data, email_var)
}
\arguments{
\item{data}{the name of the dataframe.}

\item{email_var}{string vector of email.}
}
\value{
A dataframe with proper formatted email
}
\description{
This function evaluates format of email field in the data.
Select all cases with invalid email and save a log file for further review
to the working directory named "Log - invalid_email.txt".
Replace detected invalid emails with blank in the data.
}
\examples{
data_email_validated <- epinion_email_validation(data_test, "mail")
}
