% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_sample_cleaning.R
\name{epinion_read_data}
\alias{epinion_read_data}
\title{Read data input}
\usage{
epinion_read_data(
  file,
  header = TRUE,
  sep = ";",
  fileEncoding = "UTF-16LE",
  cols = everything(),
  ...
)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{sep}{the field separator character.}

\item{fileEncoding}{character string declares the encoding used on a file.}

\item{...}{Additional arguments may apply.}
}
\value{
A dataframe of data input
}
\description{
This function reads different type of data input (xlsx, csv, rds, sav, txt, sas) into a dataframe.
It also allows users to choose only the required variables for data processing.
}
\examples{
data_test <- epinion_read_data(file = epinionSampleCleaning_example("sample_test.xlsx"))
data_test_sav <- epinion_read_data(file = epinionSampleCleaning_example("sample_test.sav"), cols = c("gruppe", "runde"))
}
