% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{epinionSampleCleaning_example}
\alias{epinionSampleCleaning_example}
\title{Get path to epinionSampleCleaning example}
\usage{
epinionSampleCleaning_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
epinionSampleCleaning comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
epinionSampleCleaning_example()
epinionSampleCleaning_example("sample_test.xlsx")
}
